LayoutTests/platform/ios-sim/storage/indexeddb/intversion-long-queue-expected.txt

CONSOLE MESSAGE: line 190: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
Test that a database is recreated correctly when an open-with-version call is queued behind both a deleteDatabase and an open-with-version call

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-long-queue.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
